kernel.mk: kernel build improvements
Author: Andrew Boie <andrew.p.boie@intel.com>
Author: Alin Dobre <alin.m.dobre@intel.com>
Author: Matt Gumbel <matthew.k.gumbel@intel.com>
Author: Daniel Leung <daniel.leung@intel.com>
Author: Lukasz Anaczkowski <lukasz.anaczkowski@intel.com>
Author: Bruce Beare <bruce.j.beare@intel.com>
Author: Igor Zamyatin <igor.zamyatin@intel.com>

The kernel build now produces 4 artifacts: the image itself, a tarball
full of modules, a tarball full of firmware, and System.map. We introduce
INSTALLED_MODULES_TARGET, INSTALLED_KERNELFW_TARGET, and INSTALLED_SYSTEM_MAP
for the latter 3 items so that they can be referenced elsewhere in the build.
If there are no modules or firmware the respective tarballs still exist
but are empty. System.map doesn't go on the device, but is extremely
useful for debugging and needs to be stored with release deliverables.

BUILT_SYSTEMIMAGE and INSTALLED_SYSTEMTARBALL_TARGET now depend on the
modules and firmware tarballs, and their contents are expanded in the
proper locations during the creation of these targets.

The kernel defconfig can now be constructed out of an arbitrary number of
fragments which are assembled by the build prior to invoking the kernel
build. This allows, for example, to have a base configuration plus some
overlays which alter it to enable/disable options based on whether a build
is debug-enabled. Users set TARGET_KERNEL_CONFIG to their base defconfig,
and can add any number of overlays to TARGET_KERNEL_CONFIG_OVERRIDES.

There is a small hack in place to fool the kernel build into using
minigzip instead of gzip when creating the bzImage.
This is because the deflate algorithm in gzip is very old compared
to zlib, and the recovery imgdiff codes create very large diffs
if the source file isn't compressed with the same version of deflate
that it has access to.

There is support for building out-of-tree kernel modules. Users will want
to set EXTERNAL_KERNEL_MODULES as a list of paths to the directories where
the modules live.

We want the kernel build process to be the same whether we are
doing a clean kernel build or an incremental one. The 'oldnoconfig'
command used updates .config.old based on the contents of .config,
but we don't want any prior state hanging around that affects how
the kernel is built. Hence delete .config.old before building each
time.

We support prebuilt kernels, modules, and firmware. Setting BUILD_KERNEL=1
will force a kernel compile, otherwise if suitable prebuilts exist in
TARGET_PREBUILT_KERNEL_DIR they will be used instead. This is desirable
for speeding up build times; the kernel build must be called as a sub-make
and the resulting images are always touched. Typical use is to have a Jenkins
job which watches for changes in the source tree and generates/commits new
prebuilts when that happens.

TODO:
- Fixup out-of-tree kernel module build to be more like other Android modules
- Remove gzip hack and patch kernel to know about minigzip
- Fix Linux kbuild Makefiles so that you can use M= and O= simultaneously
- Fix kernel_{mod|fw}_enabled to also check overrides

Change-Id: I242453ba450e5552fffa549e055663b9c2950bf0