kernel.mk: Enable dynamically-generated defconfig
Currently the kernel.mk only supports a pre-existing defconfig file. We
are considering moving to a dynamically generated defconfig and this
change enables such a thing.

I've done the following tests to verify this change:
 1) Clean MFLD full build
 2) Clean MFLD kernel build
 3) Incremental MFLD build (touch kernel source file, ensure re-build)
 4) Incremental MFLD build (touch defconfig file, ensure re-build)
 5) Clean GR build
 6) Clean GR kernel build
 7) Incremental GR build (touch source file, ensure re-build)
 8) Incremental GR build (touch defconfig file, ensure re-build)

(Please note that this change should be OK by itself even if we don't
start generating defconfigs dynamically because it increases the
flexibility of kernel.mk.)

Change-Id: I713a27cb80f33df78d9ad0875ee1fbc246e70199
Signed-off-by: Matt Gumbel <matthew.k.gumbel@intel.com>