main: Add additional subdirs with PRODUCT_CUSTOM_SUBDIRS
Define a new variable, PRODUCT_CUSTOM_SUBDIRS, that allows a product to
include additional subdirectories to be built that may have not been
detected by main.mk.

For example, when a target uses BUILD_TINY_ANDROID, main.mk generates
a minimal list of subdirectories to include in the build.  The list may be
too tiny for a given target, so PRODUCT_CUSTOM_SUBDIRS enables the
target to append to that list.

Change-Id: Iab433c61934b3b5939230225b02a4c179d75291d
Signed-off-by: Chad Versace <chad.versace@linux.intel.com>
Signed-off-by: Matt Gumbel <matthew.k.gumbel@intel.com>