MUST_REVERT kernel.mk: Fool Kbuild into using minigzip instead of gzip
We need to compress the bzImage with minigzip and not gzip.
This is because the deflate algorithm in gzip is very old compared
to zlib, and the recovery imgdiff codes create very large diffs
if the source file isn't compressed with the same version of deflate
that it has access to.  A wrapper hack script is added to the PATH
when invoking Kbuild to fool it into using minigzip, which is built
in external/zlib and staged with the host tools in the PATH.

This is a gross hack which should be reverted once we have a
cleaner solution.

Change-Id: I19b837ff7bc4d07c67cbd6342318f246842ce288