Changes for successful ics building by default toolchain
Default toolchain from ICS causes build issues in kernel.
The main reason is that compiler has been changed - it is
now built for android target, i.e.  -mandroid is on by
default and GCC documentation says that in this case -fPIC
is turned on.  Kernel sources intend to be compiled without
this option enabled so additional flag set up was made.

Issue: ANDROID-19

Change-Id: Ibdc09e3b87a60a056dd83778be124e7056e80a7c
Signed-off-by: Igor Zamyatin <igor.zamyatin@intel.com>