Automatic prebuilts conditioned by non-empty CUSTOM_BOARD
Even if the prebuilts machinery relies on the value of the
$(CUSTOM_BOARD) variable, there was no protection of including
prebuilts methods even when this variable was empty. This was causing
errors about modules not being defined correctly:
 ...
 * Each module must use a LOCAL_MODULE_TAGS in its
 * Android.mk. Possible tags declared by a module:
 ...

The fix consists in conditioning non-empty CUSTOM_BOARD both on
creating prebuilts in the auto-generated Android.mk file, and when
using incorrect prebuilts, the latter also displays an warning.

Issue: DROIDDISTRO-249
Change-Id: Ibbf5ddb37b39fb402604239c3a564fd3d1ba5936
Signed-off-by: Alin Dobre <alin.m.dobre@intel.com>