kernel.mk: Delete .config.old before building Linux
We want the kernel build process to be the same whether we are
doing a clean kernel build or an incremental one. The 'oldnoconfig'
command used updates .config.old based on the contents of .config,
but we don't want any prior state hanging around that affects how
the kernel is built. Hence delete .config.old before building each
time.

Issue: ANDROID-603
Change-Id: I33befc9dad8cffcc9d603308812751e1b4e4c41f
Signed-off-by: Andrew Boie <andrew.p.boie@intel.com>