Added support for LOCAL_CERTIFICATE in prebuilts machinery
The LOCAL_CERTIFICATE parameter is mandatory to be set for an apk package file.
Since it's a path, and is later expanded in the BUILD_PACKAGE call into the
path to the certificates, it cannot be saved into prebuilt metadata, but rather
be set in the prebuilts area.

This patch includes support for setting up this parameter, similar to the
LOCAL_PATH parameter added for modules which require installation in a separate
path.

An example of usage is:
PREBUILT.IntelWiDiApp.LOCAL_CERTIFICATE := platform

Issue: DROIDDISTRO-268
Change-Id: I11b3bb30e761ae4bba7f16ab755c7c28bb863bc5
Signed-off-by: Alin Dobre <alin.m.dobre@intel.com>