Fix for double suffixing static files
It seems that the internal ABS prebuilt boilerplate function is adding
the LOCAL_MODULE_SUFFIX module property to all modules. While this is
being handled correctly for the normal binaries and applications, it
seems that it affects the static files (like the ones in class ETC).

It's like adding a specific LOCAL_MODULE_SUFFIX:=.conf to a file.conf
module, which installs a file into /system/etc/, in which case the file
will be installed as file.conf.conf.

This patch solves this by skipping LOCAL_MODULE_SUFFIX specification for
modules that belong to the ETC class. Other classes may be added at a
later time, if necessary.

Issue: DROIDDISTRO-278
Change-Id: Ib597193655e7c7c2e219ce747b99fd80c6783dfe
Signed-off-by: Alin Dobre <alin.m.dobre@intel.com>