Enhance support for prebuilt kernel images
The way our kernel is built from source drastically increases our
overall incremental build time, since the kernel Make must be run each
time and all artifacts that depend on the kernel or modules must also
be regenerated.

This patch is mostly based on Andrew's functionality added a while
ago, however it changes a bit of its foundation: it doesn't rely on
defining or not a variable pointing to a directory that holds the
kernel prebuilts, but rather verifies the existence of these prebuilts
in a specified location, and automatically uses these prebuilts, while
the developer has always the option of forcing kernel build from
source. This functionality allows either an exclusive checkout of
either the kernel source project or the kernel prebuilts project, at
manifest level, but can also allows both kernel source and prebuilts
projects to be checked out in different directories at the same time,
keeping the flexibility of using one or the other.

The automatic prebuilts machinery uses the PREBUILT-PROJECT-* targets
for generating the binaries, and this patch also adds this target for
the kernel, in order for Jenkins to automatically generate kernel
images at each merge cycle.

Issue: DROIDDISTRO-197
Change-Id: I9dba3b4364714e131c2ce956e4cb6f13b610ac5c
Signed-off-by: Alin Dobre <alin.m.dobre@intel.com>