Fix automatic prebuilt target name (part 2)
It seems that there was a typo that passed the testing. I have named
the prebuilt target name as -kernel, where Jenkins uses the last part
of the directory name where it resides. So the correct name should end
with -linux.

Previous fix was only fixing one occurence of the PREBUILT target, now
this patch fixes the second one, too.

For: DROIDDISTRO-197
Change-Id: I843ebe50159b1ba04f5f789cfc77b0e452d40189
Signed-off-by: Alin Dobre <alin.m.dobre@intel.com>