Completely wipe out the CUSTOM_BOARD invention
Since there already are variables in the Android Build System which
hold the name of the platform, it doesn't make sense for the
CUSTOM_BOARD variable presence, especially since it's a strictly Intel
invention, it's redundant, un-useful and makes patches not suitable
for upstream submission.

This patch simply replaces all occurences of CUSTOM_BOARD with the
TARGET_BOARD_PLATFORM variable.

For: DROIDDISTRO-261
Change-Id: I85c2c746bdecb4b895a7f0cf7fd63b070336d42a
Signed-off-by: Alin Dobre <alin.m.dobre@intel.com>