kernel.mk: Don't rely on implicit population of modules into /system
We now install the modules and firmware to a staging directory instead of
directly injecting them into /system. The rules for creating a system image
now decompress the tarballs when creating the image.

Fixes a problem where the system image could be created without the modules
in it, resulting in an unusable device.

Change-Id: I8fa1c5c43a468f5bdf91c689383672d3f05cc73a
Signed-off-by: Andrew Boie <andrew.p.boie@intel.com>