Resolve apkcerts error wrt bash argument list being too long
When the BUILD_NUMBER make variable contains a longer string (tested
with 69 bytes), the command line that is generated by the ABS is
longer than 128k, which is, at this moment, a bash maximum. This leads
to the following error:

make: execvp: /bin/bash: Argument list too long
make: *** [<product-out>/obj/PACKAGING/apkcerts_intermediates/<some-name>.txt] Error 127

The cause for this is that the command line contains all the
name="...", certificate="..." and private_key="..." strings echos,
along with the redirect file, multiplied by many times, as many as the
packages included in the compiled product.

The solution from this patch keeps only doing echos for the strict
information, and the rest of the echoing is done in a bash for loop.

Change-Id: I7c57b6c319cc1ed78e279479002b4801175405cb
Signed-off-by: Alin Dobre <alin.m.dobre@intel.com>