Fix the SDK build.
In the top-level Makefile, a couple of chmod commands were recently
added to a recipe, but the targets of the chmod commands may not
exist in all cases, e.g. when building the SDK. If a target of either
of the chmod command is not there, the chmod command fails, and the
build process does, too.

This change makes 'make' ignore the exit status from the chmods.

Change-Id: I1620f6e6f77dbdbdef78239197eda949ead28471
Signed-off-by: Tom Keel <thomas.keel@intel.com>