select.mk: make path to ccache externally visible as CCACHE_BIN
Some other Makefiles can make use of this information if they
need to invoke it directly. We change the variable to uppercase
to follow convention of Make variables that are intended to be
accessed outside the scope of the current file.

The immediate use-cases for this are extensions to the build system
to build the Linux kernel in a sub-make, and some of the QEMU
Makefiles which call the compiler in an unusual way.

Bug: ADEV-129
Category: AOSP Improvement
Domain: AOSP-build
Severity: Medium
Origin: Internal
Upstream-Candiate: Yes

Change-Id: I31dbff712f71bed0db85f8de16ad005d8d1c9578
Signed-off-by: Andrew Boie <andrew.p.boie@intel.com>