build: Conditionally include iago Image creation Makefiles
Iago is the next-generation Diskinstaller for Legacy BIOS and EFI-
based Intel devices. It is a replacement for the existing
Diskinstaller program.

This file contains directives which cannot go in an Android.mk
and needs to be after they all have been included, similar to the
old Diskinstaller ramdisk directives.

Additionally include the Iago provision.mk file, which needs
to be included after the rules for the target-files-package.
These Makefile directives need to live in the same repo as
the rest of the Iago code and are hence not included in a
vendor/*/build/tasks directory.

Bug: ADEV-129
Category: Device Enablement
Domain: AOSP-build, INSTALLER-iago
Severity: High
Origin: Internal
Upstream-candidate: Yes, but only when we are ready to upstream
                    the bootable/iago project

Change-Id: I4c07f13a857f19052e15c562805f8f64452e4c3d
Signed-off-by: Andrew P. Boie <andrew.p.boie@intel.com>

Conflicts:
	core/Makefile