sign-target-files-apks: optionally build boot images
ota_from_target_files will build the boot images itself unless
it finds prebuilt copies in the TFP under BOOTABLE_IMAGES/.

We'd like to keep it so that the production keys are only required
during the sign_target_files_apks step. So the user can specify to
have the boot images created, overriding mkbootimg_args to supply
a path to the production key.

Change-Id: Ifd83bd9087fa050a27ec7f882093ae5a01940040
For: AXIA-2386
Signed-off-by: Andrew Boie <andrew.p.boie@intel.com>