Build bison before compiling kernel and external modules from source
The kernel perf tool requires bison to build. Since bison is not
a hard requirement for building kernel and external modules, there are
times where it is using the host system's bison. Depending on when
bison is built, some may be built by host's bison, some may be built
by Android's bison.

This patch forces kernel, if building from source, to depend on
Android's bison. This causes all needed modules to be built with
Android's bison.

Change-Id: I792e531902d1451ed8eb718fbc0cc8c70c1fad8d
Signed-off-by: Daniel Leung <daniel.leung@intel.com>