Fix EXTRA_KERNEL_MODULES build when more than one is defined
When the list is e.g. "a b" (or in this case "iwlwifi iwlwifi" in my
tree, which seems like a separate bug), this foreach expression
expands to "make a make b" and the build fails because it doesn't know
how to build "make" as a target.  There's no point in spawning a
separate make for those anyway that I can see.

Change-Id: Id1077435342bae470e95ea1576dd849215a0b3b3
Signed-off-by: Andy Ross <andrew.j.ross@intel.com>