Makefile: add support for copyleft code on the device itself
With space being as cheap as it is, some projects may want to
simply include all the GPL source code on the device itself
in an archive, so that if users really want to see it they can
do so using a filesystem explorer app.

Use of this feature will increase incremental rebuild times as
there is no robust way to detect if the source code changes;
the rules for all these tarballs are declared .PHONY.
However, we have some logic so that things that depend on these
tarballs (like INSTALLED_SYSTEMIMAGE_TARGET) aren't updated
unless something actually changes.

Issue: ADEV-359
Change-Id: Ibb5b43e3751c574676229dd4fba7a5f5a27124ab
Category: AOSP Improvement
Domain: AOSP-build
Upstream-candidate: Yes
Signed-off-by: Andrew Boie <andrew.p.boie@intel.com>