Enable build of kernel compat modules
Kernel compat modules are Linux kernel drivers that are taken from
a kernel tree and backported to an old kernel. The driver
is not integrated into the old kernel tree, but version controlled
and build apart. A customized set of makefiles are bundled with
the driver which enables the driver to be built, using the old
kernel's makefiles and Kconfig files.
This patch introduces an Android-like build interface that needs
to be used from the Android.mk file of each compat module.
The make variables set by the caller are:
  LOCAL_MODULE - name of the module
    Used for defining unique make rules and useful printouts
  LOCAL_PATH
    Path to module, relative to root build dir
  LOCAL_KERNEL_COMPAT_DEFCONFIG
    Name of the module-local defconfig file used to configure
    the compat build. The full name passed to the kernel makefile
    is defconfig-$(LOCAL_KERNEL_COMPAT_DEFCONFIG)
  COMPAT_PRIVATE_$(LOCAL_MODULE)_PREINSTALL:
  COMPAT_PRIVATE_$(LOCAL_MODULE)_POSTINSTALL:
    Defines a set of commands that will be run before/after the
    module is installed.
    This macro should be defined using define/endef, to ensure
    that it uses deferred evaluation. And yes, the macro may
    contain multiple lines.
    Two parameters may be used inside the commands:
    $(1): Path to where module is installed
    $(2): Path to where the module is built from

Rationale for not upstreaming: these are fairly Intel-specific
build modifications.

Change-Id: I2e58e4622b8927b22fcea44ea797c5976b350123
Depends-Change-Id: Ifd17ae3eca46f64f52d0e684b6bb320e335b7cfd
Category: Device Enablement
Domain: Build system
Upstream-Candidate: No
For: APDEV-648
Signed-off-by: Hans Odeberg <hans.odeberg@intel.com>
[apboie: removed legacy method]
Signed-off-by: Andrew Boie <andrew.p.boie@intel.com>