kernel: add kernel sources to prebuilt archives
For GPL compliance we need to be able to publish the exact kernel
sources used to build the kernel that is on the device. Use of
prebuilt kernels complicates this because the source checkout
(if it even exists) may not be the same code as what was compiled
in the prebuilt archive.

Include a tarball of the kernel sources inside the prebuilt
kernel archives. Publish a new variable
$(INSTALLED_KERNEL_SOURCE_TARBALL) so that other parts of the
build system can reference it if needed, independent of whether
prebuilts are used or not.

Introduce a new build variable ALL_EXTRA_SOURCE_TARBALLS that
specifies a list of extra archives to include in any mechanism
the collects sources for GPL/MPL/LGPL compliance. Amend
collect_gpl_sources.mk to use this.

If for some reasons additional kernel code not in
TARGET_KERNEL_SOURCE or modules is needed, directories
containing it can be specified with TARGET_EXTRA_KERNEL_SOURCE.

Amend the rules for building out-of-tree modules such that
MODULE_LICENSE_GPL and friends is checked for, and add
such sources to the kernel source archive.

Issue: ADEV-359
Category: AOSP Improvement
Domain: AOSP-build
Upstream-candidate: Only if we upstream the base kernel build rules
Change-Id: Ib67b7b8d335d152cee5100b7e9e2a107403bf30f
Signed-off-by: Andrew Boie <andrew.p.boie@intel.com>