Only rebuild compat kernel modules when needed
Compat kernel modules are configured (make defconfig) at each
build. This causes the .config file to be touched, which then
during build causes the compat build's autoconf.h file to
be touched. And as this file is included by each source file,
the entire module would be rebuilt from scratch, every time.
In the normal kernel build the dependency to autoconf.h is
filtered out from the dependency list, and replaced with
dependencies to files in include/config. However, as the
compat module's autoconf.h file is stored in a different path,
it is not filtered out.
The solution to the problem is to generate a temporary .config.tmp
file during the compat module configuration. The temp file is
copied to .config only if it differs from the existing .config file.
This solution will cause the entire compat module to be rebuilt if
there is a configuration change. The full, fine-grained rebuild
resolution of the kernel build system is thus not achieved.

Issue: APDEV-673
Change-Id: I96f6b02c6793d0f4fb3785dddc35eae630034875
Category: AOSP Improvement
Domain: AOSP-build
Upstream-candidate: Only if we upstream the base kernel build rules
Signed-off-by: Hans Odeberg <hans.odeberg@intel.com>