Enable parallel build of external kernel modules
Kernel modules are built independently of each other. A first step
includes copying all source code to a location in the output tree.
If one module is located in a subdirectory to another module, the
two copy operations may collide and cause the build to fail. This
patch ensures separate target directories for all modules, by
including the module name in the path.

As the module in the subdirectory may try to include header files
above its own top directory, the variable LOCAL_C_INCLUDES is
used to identify additional header file directories.

Not considered for upstream as it modifies and depends on
Intel-specific additions.

Issue: APDEV-1300
Change-Id: Ibbc1a2229c3253c0278f2fa6fec59626e6cc06ca
Depends-Change-Id: Ie758d18031dac2b09eaad7b7579a94c4c1e45ffc
Category: Device Enablement
Domain: Build system
Upstream-Candidate: No
Signed-off-by: Hans Odeberg <hans.odeberg@intel.com>